feat: add deepcopy support for eLinear and InvariantConstraint classes
- Implemented __deepcopy__ method in eLinear to maintain shared representations and cached tensors.
- Added __deepcopy__ method in InvariantConstraint to ensure proper copying of internal states.
- Refactored module_memory and bytes_to_mb functions in utils.py for clarity and organization.
